requirements.txt
torch
torchvision
tqdm
matplotlib
openai
requests
python-dotenv
pandas
scikit-learn
diffusers
transformers
